SOFT-3992 proposal: move Projection/Variant to Projection/Preset and update refs
